feat: dto Long -> BigDemical 타입으로 변경

+++ backend/coinflow-ws-gateway/src/main/java/com/coinflow/ws/dto/TickDto.java
@@ -5,8 +5,8 @@ import lombok.Builder;
 
 @Builder
 public record TickDto(
-        String symbol,
-        BigDecimal price,
-        Long volume,
-        Long eventTime) {
+                String symbol,
+                BigDecimal price,
+                BigDecimal volume,
+                Long eventTime) {
 }
